requirements.txt
nidaqmx
pyqt5
pyqtgraph
pydantic
numpy==1.26.4
paho-mqtt